% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/HW6-package.R
\docType{package}
\name{HW6-package}
\alias{HW6-package}
\title{Homework6}
\description{
The package contains 3 functions of Homework 6 of Advanced
  Data Analysis Software Development in R.
  1. Function mode() - determines the most frequently occurring value in an integer vector
  2. Function simplify2array()  - simplify the list of numeric vectors to numeric matrix
  3. Function perms() generates all the possible permutations of the set {1,2,...,n} (for some given n)
}

