% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mode}
\alias{mode}
\title{Mode}
\usage{
mode(x)
}
\arguments{
\item{x}{- integer vector}
}
\value{
Integer, containing the most frequently occurring value in the input vector
}
\description{
function mode() determines the most frequently occurring value in an integer vector (mode).
}
\examples{
mode(c (5,8,3,9,3,0,1,5,4,5,9))
mode(c (1))
mode(1:10)
}

