% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{perms}
\alias{perms}
\title{perms}
\usage{
perms(n)
}
\arguments{
\item{n}{- positive integer}
}
\value{
a matrix with n! rows and n columns
}
\description{
Function perms() generates all the possible permutations of the set {1,2,...,n} (for some given n)
}
\examples{
perms(3)
perms(1)
}

